+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixing samsung routines and modes
 - Fixing wrong orientation on start if holding the phone wrong. Now uses the orientation from settings
 -  Try to fix usb errors with AAwireless Dongles
+- Added Audio Mixer to mix different audio tracks, thanks to @jeffdapaz for the idea
 
 ### v.2.3.0
 - Added some new buttons for keymap